+++ include/render/allocator/shm.h
@@ -1,8 +1,8 @@
-#ifndef RENDER_SHM_ALLOCATOR_H
-#define RENDER_SHM_ALLOCATOR_H
+#ifndef RENDER_ALLOCATOR_SHM_H
+#define RENDER_ALLOCATOR_SHM_H
 
 #include <wlr/types/wlr_buffer.h>
-#include "render/allocator.h"
+#include "render/allocator/allocator.h"
 
 struct wlr_shm_buffer {
 	struct wlr_buffer base;
