+++ render/allocator/shm.c
@@ -5,7 +5,7 @@
 #include <unistd.h>
 #include <wlr/util/log.h>
 #include "render/pixel_format.h"
-#include "render/shm_allocator.h"
+#include "render/allocator/shm.h"
 #include "util/shm.h"
 
 static const struct wlr_buffer_impl buffer_impl;
