+++ render/meson.build
@@ -6,20 +6,13 @@ elif 'auto' in renderers and get_option('auto_features').disabled()
 endif
 
 wlr_files += files(
-	'allocator.c',
 	'dmabuf.c',
 	'drm_format_set.c',
-	'gbm_allocator.c',
 	'pixel_format.c',
-	'shm_allocator.c',
 	'swapchain.c',
 	'wlr_renderer.c',
 	'wlr_texture.c',
-	'drm_dumb_allocator.c',
 )
-
-has = cc.has_function('gbm_bo_get_fd_for_plane', dependencies: [gbm])
-add_project_arguments('-DHAS_GBM_BO_GET_FD_FOR_PLANE=@0@'.format(has.to_int()), language: 'c')
 
 egl = dependency('egl', required: 'gles2' in renderers)
 if egl.found()
@@ -32,3 +25,5 @@ if 'gles2' in renderers or 'auto' in renderers
 endif
 
 subdir('pixman')
+
+subdir('allocator')
